Website update 001 - Initial pass of entire site done.  (#9)
* Initial clean up check in. Getting theme and css overides set

Getting this theme back into the way it intended to be used. Overiding the theme with more specificity in css. Removing the theme .scss files.

* Clean up and Progress

I've now got basically everything in the design. Still needs some style work. Will need a pass at mobile still. BUT this is everything in the design and is getting very close.

* Update: Redo of the Spec and System Requirements

Basically an overhaul of the site.. Getting evertying stubbed in to work on desktop.

* Website Update: Final Clean up and Mobile

I have restructured the home page yet again. I cleaned up styles and took a pass at responsive breakpoints.

All links should either go viewing the manual or the GitHub project page.