Fixing footer buttons reordering main site
I made it so that the social buttons in the footer now are stacked. They were getting wonky in Safari when you resized your browser. That is fixed now.

Based on the slack #design channel I reorderd the sections to this:

The basic flow down the page that is being proposed is the following:

Hero section (d/l’s with giant screen shot)
Digital Beauty
Sounds
Surge is back and Open Source (giving props to kurasu for the OSS :heart:)
Specifications
Contribution stuff
System req’s