Convert to jpeg and fix more responsive issues (#121)
* converting website .png to .jpg where we can

Making most of the image assets in the site .jpg so that we load faster. Also removing the references to unused files. Updated the  stylesheet to reference the new .jpg files.

* Fixing responsive issues again

Made it so we are responsive down to 299px. Beyond that and we still have problems. Not sure it matters though. More work could be done to fix it up. Given the current viewport width of devices in the wild Im inclined to leave it all alone.