Header layout changes (#188)
* Initial Navbar Checkin

First pass at nabar.

* slight tweak to navbar styles

* reading nav links from a list (mostly)

For the home link and the manual I have opted to just and roll the links (home because its just the base site URL and manual because of the logic switching on selection).

The rest of the links in the nabar are supplied in a _data folder that contains a surge_navItem.yaml file.

In that file you define Surge Pages. Where Surge Pages has docs that have title and URL.

The title is what displays in the navlink. The URL is ummmm the URL which is either relative to the site or an internet addy.

* fixing hero image now

At large screen sizes the hero image was getting clipped off the top and the bottom of the image.

Fixing this by putting margin on the image and setting

* fixing navlinks and small style tweaks

* sorting out the hover state on nav links

also updated some formatting and added a couple of color variables to the theme.