Hero section update (#190)
* Changes to hero information

Removing some of the information density. Ive now gone far enough that I dont want to loose it.

* Adding Animation Classes

Hooking up animation for some things. changing the hero more so that once you load the body we update the button and link on the hero based on the OS we detect.

**TODO**
fix the loading so that the lable change is less obvious.

* More changes to hero section

Updating hero copy and making mobile better.

* fixing up further mobile issues.

* removing comment