Cleaning up hero animation
changing it so that the whole page no longer has the animation happen onLoad. Now only the text elements in the hero animate in. Since the size of the hero is 100% VH & VW it will take up the users viewport. No need for the animation to run on the  entire body of the site.