fixing scrolling to sections in the manual
The header is fixed, the sections were getting tucked up underneath the header when clicking on a section in the manual TOC.

This fixes that issue by adding padding to the top of the scroll.

Support for this new(ish) CSS feature seems pretty good. It hits all majors except for IE 11.....

https://caniuse.com/?search=scroll-padding-top