Slight update to instructions
I set up a new machine tonight and attempted to follow the instructions outlined here. I ran into issues with installation of gemfiles. I was able to solve them by following a fairly straightforward tutorial for macOS. It also required a slight reordering of the commands to get things up and running.